backend/auth_service/requirements.txt
fastapi
uvicorn
psycopg2-binary
sqlalchemy
mangum
PyJWT
python-dotenv